requirements.txt
flask==3.0.2
waitress==3.0.0
python-dotenv==1.0.1
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
scikit-learn>=1.5.0
pytest>=8.0.0
requests>=2.31.0